Convert to React.createContext API
