+++ packages/styletron-react-core/package.json
@@ -30,6 +30,7 @@
     "prepublish": "npm run build"
   },
   "dependencies": {
+    "create-react-context": "^0.2.2",
     "prop-types": "^15.6.0"
   },
   "peerDependencies": {
