+++ packages/styletron-react-core/src/__tests__/tests.browser.js
@@ -5,30 +5,179 @@ import Enzyme from "enzyme";
 import Adapter from "enzyme-adapter-react-16";
 import * as React from "react";
 
-import {withWrapper, withStyle, Provider} from "../index.js";
+import {
+  withWrapper,
+  withStyle,
+  withStyleDeep,
+  withTransform,
+  Provider,
+} from "../index.js";
 import {styled} from "./utils/custom-styled.js";
 
 Enzyme.configure({adapter: new Adapter()});
 
-test("Provider", t => {
-  t.plan(1);
-  const styletron = {};
-
-  const MockComponent = (props, context) => {
-    t.equal(
-      context.styletron,
-      styletron,
-      "styletron instance override provided",
-    );
-    return <div />;
-  };
-  MockComponent.contextTypes = {styletron: () => {}};
-
+test("styled (static)", t => {
+  t.plan(3);
+  const style = {size: 1};
+  const Widget = styled("div", style);
   Enzyme.mount(
-    <Provider value={styletron}>
-      <MockComponent />
+    <Provider
+      value={{
+        renderStyle: x => {
+          t.deepEqual(x, style);
+        },
+      }}
+    >
+      <Widget />
     </Provider>,
   );
+  const wrapper = Enzyme.mount(
+    <Provider
+      value={{
+        renderStyle: () => "bar",
+      }}
+    >
+      <Widget className="foo" />
+    </Provider>,
+  );
+  const divs = wrapper.find("div");
+  t.equal(divs.length, 1, "single div rendered");
+  t.ok(divs.hasClass("foo bar"), "explicit and generated class names merged");
+  t.end();
+});
+
+test("styled (dynamic)", t => {
+  t.plan(1);
+  const Widget = styled("div", (props: {foo: boolean}) => ({
+    size: props.foo ? 1 : 2,
+  }));
+
+  Enzyme.mount(
+    <Provider
+      value={{
+        renderStyle: x => {
+          t.deepEqual(x, {size: 1});
+        },
+      }}
+    >
+      <Widget foo={true} />
+    </Provider>,
+  );
+  t.end();
+});
+
+test("withStyle (static)", t => {
+  t.plan(1);
+  const Widget = styled("div", {size: 1, color: "red"});
+  const SuperWidget = withStyle(Widget, {
+    color: "blue",
+    shape: "circle",
+  });
+  Enzyme.mount(
+    <Provider
+      value={{
+        renderStyle: x => {
+          t.deepEqual(x, {size: 1, color: "blue", shape: "circle"});
+        },
+      }}
+    >
+      <SuperWidget />
+    </Provider>,
+  );
+  t.end();
+});
+
+test("withStyle (dynamic)", t => {
+  t.plan(1);
+  const Widget = styled("div", {color: "red", shape: "square"});
+  const SuperWidget = withStyle(Widget, props => ({
+    shape: props.$round ? "circle" : "square",
+    size: 1,
+  }));
+  Enzyme.mount(
+    <Provider
+      value={{
+        renderStyle: x => {
+          t.deepEqual(x, {color: "red", shape: "circle", size: 1});
+        },
+      }}
+    >
+      <SuperWidget $round={true} />
+    </Provider>,
+  );
+  t.end();
+});
+
+test("withStyleDeep (static)", t => {
+  t.plan(1);
+  const Widget = styled("div", {size: 1, color: "red", velocity: {speed: 1}});
+  const SuperWidget = withStyleDeep(Widget, {
+    color: "blue",
+    velocity: {direction: 1},
+  });
+  Enzyme.mount(
+    <Provider
+      value={{
+        renderStyle: x => {
+          t.deepEqual(x, {
+            size: 1,
+            color: "blue",
+            velocity: {speed: 1, direction: 1},
+          });
+        },
+      }}
+    >
+      <SuperWidget />
+    </Provider>,
+  );
+  t.end();
+});
+
+test("withStyleDeep (dynamic)", t => {
+  t.plan(1);
+  const Widget = styled("div", {size: 1, color: "red", velocity: {speed: 1}});
+  const SuperWidget = withStyleDeep(Widget, props => ({
+    shape: props.$round ? "circle" : "square",
+    velocity: {direction: 1},
+  }));
+  Enzyme.mount(
+    <Provider
+      value={{
+        renderStyle: x => {
+          t.deepEqual(x, {
+            color: "red",
+            shape: "circle",
+            size: 1,
+            velocity: {direction: 1, speed: 1},
+          });
+        },
+      }}
+    >
+      <SuperWidget $round={true} />
+    </Provider>,
+  );
+  t.end();
+});
+
+test("withTransform", t => {
+  t.plan(1);
+  const Widget = styled("div", {color: "red", shape: "square"});
+  const SuperWidget = withTransform(Widget, (style, props) => ({
+    ...style,
+    shape: props.$round ? "circle" : "square",
+  }));
+  Enzyme.mount(
+    <Provider
+      value={{
+        renderStyle: x => {
+          t.deepEqual(x, {color: "red", shape: "circle"});
+        },
+      }}
+    >
+      <SuperWidget $round={true} />
+    </Provider>,
+  );
+  t.end();
 });
 
 test("$as works", t => {
@@ -38,22 +187,26 @@ test("$as works", t => {
     t.equal(props.className, "foo", "styled class is passed");
     return <div />;
   };
-  Enzyme.mount(<Widget $as={MockComponent} />, {
-    context: {
-      styletron: {
+  Enzyme.mount(
+    <Provider
+      value={{
         renderStyle: () => {
           return "foo";
         },
-      },
-    },
-  });
-  const wrapper = Enzyme.mount(<Widget $as="span" />, {
-    context: {
-      styletron: {
+      }}
+    >
+      <Widget $as={MockComponent} />
+    </Provider>,
+  );
+  const wrapper = Enzyme.mount(
+    <Provider
+      value={{
         renderStyle: () => {},
-      },
-    },
-  });
+      }}
+    >
+      <Widget $as="span" />
+    </Provider>,
+  );
   t.equal(wrapper.find("span").length, 1, "span rendered");
   t.end();
 });
@@ -77,15 +230,17 @@ test("$-prefixed props not passed", t => {
 
   const Widget = styled(InnerComponent, {size: 1});
 
-  Enzyme.mount(<Widget $foo="foo" $baz="baz" data-bar="bar" />, {
-    context: {
-      styletron: {
+  Enzyme.mount(
+    <Provider
+      value={{
         renderStyle: () => {
           return "styleclass";
         },
-      },
-    },
-  });
+      }}
+    >
+      <Widget $foo="foo" $baz="baz" data-bar="bar" />
+    </Provider>,
+  );
 });
 
 test("$ref", t => {
